+++ TextBasedGame/GameController.h
@@ -9,7 +9,7 @@
 #include <ctime>
 #include <stdio.h>
 #include <time.h>
-#include "Settings.h";
+#include "Settings.h"
 //This adds in a function for me so I can do delay the console
 #if defined(__WIN32__) || defined(_WIN32) || defined(WIN32) || defined(__WINDOWS__) || defined(__TOS_WIN__)
 
@@ -18,7 +18,7 @@
 inline void delay(double seconds)
 {
 
-	Sleep((seconds * 1000));
+	Sleep(((DWORD)(seconds * 1000)));
 }
 
 #else  /* presume POSIX */
